+++ Source/ShadowMap/ShadowMapService.cpp
@@ -106,8 +106,8 @@ void ShadowMapService::CreateShadowRenderSystem(const std::string& renderer_name
         m_renderer.lock()->SetAssociatedCamera(m_sunLightCamera);
     }
 
-    m_shadowMapDimensionBiasDensity[0] = m_configuration->ShadowMapDimension().m_width;
-    m_shadowMapDimensionBiasDensity[1] = m_configuration->ShadowMapDimension().m_height;
+    m_shadowMapDimensionBiasDensity[0] = static_cast<float>(m_configuration->ShadowMapDimension().m_width);
+    m_shadowMapDimensionBiasDensity[1] = static_cast<float>(m_configuration->ShadowMapDimension().m_height);
     m_shadowMapDimensionBiasDensity[2] = m_configuration->ShadowMapDepthBias();
     m_shadowMapDimensionBiasDensity[3] = m_configuration->ShadowMapDensity();
 }
